TEIID-2702 using an Integer for batch size

+++ connectors/connector-google/src/main/java/org/teiid/resource/adapter/google/SpreadsheetConnectionImpl.java
@@ -134,7 +134,7 @@ public class SpreadsheetConnectionImpl extends BasicConnection implements Google
 			String worksheetName, String query, 
 			 Integer offset, Integer limit, int batchSize) {
 		
-		return dataProtocol.executeQuery(config.getSpreadsheetName(), worksheetName, query, Math.min(batchSize, (int)config.getBatchSize()), 
+		return dataProtocol.executeQuery(config.getSpreadsheetName(), worksheetName, query, Math.min(batchSize, config.getBatchSize()), 
 				offset, limit);
 	}
